+++ PUP-LIB/pup-global.lua
@@ -1,0 +1,66 @@
+-- Default States
+Master_State = "Idle"
+Pet_State = "Idle"
+
+-- Constants in case we decide to change names down the road, will be much easier
+const_dd = "DD"
+const_tank = "TANK"
+const_mage = "MAGE"
+const_PetModeCycle = "PetModeCycle"
+const_PetStyleCycle = "PetStyleCycle"
+const_stateIdle = "Idle"
+const_stateEngaged = "Engaged"
+const_stateOverdrive = "Overdrive"
+const_petOnly = "Pet Only"
+const_masterOnly = "Master Only"
+
+currentTargetedMonster = 0
+previousTargetedMonster = 0
+
+--- SKILLCHAIN TABLE
+SC = {}
+SC["Valoredge Frame"] = {}
+SC["Sharpshot Frame"] = {}
+SC["Harlequin Frame"] = {}
+SC["Stormwaker Frame"] = {}
+
+SC["Valoredge Frame"]["Stringing Pummel"] = "String Shredder"
+SC["Valoredge Frame"]["Victory Smite"] = "String Shredder"
+SC["Valoredge Frame"]["Shijin Spiral"] = "Bone Crusher"
+SC["Valoredge Frame"]["Howling Fist"] = "String Shredder"
+
+SC["Sharpshot Frame"]["Stringing Pummel"] = "Armor Shatterer"
+SC["Sharpshot Frame"]["Victory Smite"] = "Armor Shatterer"
+SC["Sharpshot Frame"]["Shijin Spiral"] = "Armor Piercer"
+SC["Sharpshot Frame"]["Howling Fist"] = "Arcuballista"
+SC["Sharpshot Frame"]["Dragon Kick"] = "Armor Shatterer"
+SC["Sharpshot Frame"]["One Inch Punch"] = "Daze"
+SC["Sharpshot Frame"]["Spinning Attack"] = "Armor Shatterer"
+SC["Sharpshot Frame"]["Base"] = "Arcuballista"
+
+SC["Harlequin Frame"]["Stringing Pummel"] = "Slapstick"
+SC["Harlequin Frame"]["Victory Smite"] = "Magic Mortar"
+SC["Harlequin Frame"]["Shijin Spiral"] = "Slapstick"
+SC["Harlequin Frame"]["Howling Fist"] = "Knockout"
+
+SC["Stormwaker Frame"]["Stringing Pummel"] = "Slapstick"
+SC["Stormwaker Frame"]["Victory Smite"] = "Magic Mortar"
+SC["Stormwaker Frame"]["Shijin Spiral"] = "Slapstick"
+SC["Stormwaker Frame"]["Howling Fist"] = "Knockout"
+
+-- Puppet Weaponskill Modifiers
+Modifier = {}
+
+Modifier["String Shredder"] = "VIT"
+Modifier["Bone Crusher"] = "VIT"
+Modifier["Armor Shatterer"] = "DEX"
+Modifier["Armor Piercer"] = "DEX"
+Modifier["Arcuballista"] = "DEXFTP"
+Modifier["Daze"] = "DEXFTP"
+Modifier["Slapstick"] = "DEX"
+Modifier["Knockout"] = "AGI"
+
+-- Pet Weapon Skills we are checking against in job_pet_aftercast
+AutomatonWeaponSkills = T {"Slapstick", "Knockout", "Magic Mortar", "Chimera Ripper", "String Clipper",
+                           "Cannibal Blade", "Bone Crusher", "String Shredder", "Arcuballista", "Daze", "Armor Piercer",
+                           "Armor Shatterer"}
