+++ PUP-LIB/pup-utility.lua
@@ -1,0 +1,138 @@
+-- Prints to the screen in a certain format
+function msg(str)
+    send_command("@input /echo *-*-*-* " .. str .. " *-*-*-*")
+end
+
+-- Traverses a table to see if it contains the given element
+function table.contains(table, element)
+    for _, value in pairs(table) do
+        if string.lower(value) == string.lower(element) then
+            return true
+        end
+    end
+    return false
+end
+
+-- Takes a condition and returns a given value based on if it is true or false
+function ternary(cond, T, F)
+    if cond then
+        return T
+    else
+        return F
+    end
+end
+
+-- Dump the contents of a table
+function dump(o)
+    if type(o) == "table" then
+        local s = "{ "
+        for k, v in pairs(o) do
+            if type(k) ~= "number" then
+                k = '"' .. k .. '"'
+            end
+
+            if type(k) ~= "number" and (k:contains("_raw") or k:contains("_data")) then
+                s = s .. "[" .. k .. "] = " .. 'binaray' .. ",\n"
+            else
+                s = s .. "[" .. k .. "] = " .. dump(v) .. ",\n"
+            end
+        end
+        return s .. "} "
+    else
+        return tostring(o)
+    end
+end
+
+-- Attempts to determine the Puppet Mode and Style
+function determinePuppetType()
+    local head = pet.head
+    local frame = pet.frame
+
+    local ValHead = "Valoredge Head"
+    local ValFrame = "Valoredge Frame"
+
+    local HarHead = "Harlequin Head"
+    local HarFrame = "Harlequin Frame"
+
+    local SharpHead = "Sharpshot Head"
+    local SharpFrame = "Sharpshot Frame"
+
+    local StormHead = "Stormwaker Head"
+    local StormFrame = "Stormwaker Frame"
+
+    local SoulHead = "Soulsoother Head"
+    local SpiritHead = "Spiritreaver Head"
+
+    -- This is based mostly off of the frames from String Theory (most are present, may be missing some)
+    -- https://www.bg-wiki.com/bg/String_Theory#Automaton_Frame_Setups
+
+    -- Determine Head first, then further determine by body and attachments
+    if head == HarHead then -- Harlequin Predictions
+        if frame == HarFrame and (pet.attachments.strobe == true or pet.attachments.flashbulb == true) then -- Magic Tank
+            handle_set({const_PetModeCycle, const_tank})
+            handle_set({const_PetStyleCycle, "MAGIC"})
+        elseif frame == HarFrame then -- Default
+            handle_set({const_PetModeCycle, const_dd})
+            handle_set({const_PetStyleCycle, "NORMAL"})
+        else
+            msg("Unable to determine Mode/Style for Puppet Head: (" .. head .. ") Puppet Frame: (" .. frame .. ")")
+        end
+    elseif head == ValHead then -- Valoredge Predictions
+        if frame == SharpFrame then
+            if (pet.attachments.strobe == true or pet.attachments.flashbulb == true) then -- DD Tank
+                handle_set({const_PetModeCycle, const_tank})
+                handle_set({const_PetStyleCycle, const_dd})
+            else -- Default
+                handle_set({const_PetModeCycle, const_dd})
+                handle_set({const_PetStyleCycle, "NORMAL"})
+            end
+        elseif frame == ValFrame then -- Default Standard Tank
+            if pet.attachments.inhibitor == true or pet.attachments.attuner == true and
+                (not pet.attachments.strobe or not pet.attachments['strobe ii']) then -- Bone Slayer
+                handle_set({const_PetModeCycle, const_dd})
+                handle_set({const_PetStyleCycle, "BONE"})
+            else -- Standard Tank
+                handle_set({const_PetModeCycle, const_tank})
+                handle_set({const_PetStyleCycle, "NORMAL"})
+            end
+        else
+            msg("Unable to determine Mode/Style for Puppet Head: (" .. head .. ") Puppet Frame: (" .. frame .. ")")
+        end
+    elseif head == SharpHead then -- Sharpshooter Prediction
+        if frame == SharpFrame then -- SPAM DD
+            if (pet.attachments.inhibitor == true or pet.attachments["inhibitor II"] == true) then
+                handle_set({const_PetModeCycle, const_dd})
+                handle_set({const_PetStyleCycle, "NORMAL"})
+            else
+                handle_set({const_PetModeCycle, const_dd})
+                handle_set({const_PetStyleCycle, "SPAM"})
+            end
+        else
+            msg("Unable to determine Mode/Style for Puppet Head: (" .. head .. ") Puppet Frame: (" .. frame .. ")")
+        end
+    elseif head == StormHead then -- Stormwaker Prediction
+        if frame == StormFrame then -- RDM
+            handle_set({const_PetModeCycle, const_mage})
+            handle_set({const_PetStyleCycle, "SUPPORT"})
+        else
+            msg("Unable to determine Mode/Style for Puppet Head: (" .. head .. ") Puppet Frame: (" .. frame .. ")")
+        end
+    elseif head == SoulHead then -- Soulsoother Prediction
+        if frame == StormFrame then -- WHM
+            handle_set({const_PetModeCycle, const_mage})
+            handle_set({const_PetStyleCycle, "HEAL"})
+        elseif frame == ValFrame then -- Turtle Tank
+            handle_set({const_PetModeCycle, const_tank})
+            handle_set({const_PetStyleCycle, "NORMAL"})
+        else
+            msg("Unable to determine Mode/Style for Puppet Head: (" .. head .. ") Puppet Frame: (" .. frame .. ")")
+        end
+    elseif head == SpiritHead then -- Spiritweaver Prediction
+        if frame == StormFrame then -- BLM
+            handle_set({const_PetModeCycle, const_mage})
+            handle_set({const_PetStyleCycle, const_dd})
+        else
+            msg("Unable to determine Mode/Style for Puppet Head: (" .. head .. ") Puppet Frame: (" .. frame .. ")")
+        end
+    end
+end
