+++ PUP.lua
@@ -16,10 +16,8 @@
     Contributors: Xilkk, Byrne, Blackhalo714
 
     ASCII Art Generator: http://www.network-science.de/ascii/
-    
-]]
 
--- Initialization function for this job file.
+]] -- Initialization function for this job file.
 -- IMPORTANT: Make sure to also get the Mote-Include.lua file (and its supplementary files) to go with this.
 function get_sets()
     mote_include_version = 2
@@ -33,7 +31,7 @@ function user_setup()
 
     --[[
         F9 - Cycle Offense Mode (the offensive half of all 'hybrid' melee modes).
-        
+
         These are for when you are fighting with or without Pet
         When you are IDLE and Pet is ENGAGED that is handled by the Idle Sets
     ]]
@@ -41,7 +39,7 @@ function user_setup()
 
     --[[
         Ctrl-F9 - Cycle Hybrid Mode (the defensive half of all 'hybrid' melee modes).
-        
+
         Used when you are Engaged with Pet
         Used when you are Idle and Pet is Engaged
     ]]
@@ -49,7 +47,7 @@ function user_setup()
 
     --[[
         Alt-F12 - Turns off any emergency mode
-        
+
         Ctrl-F10 - Cycle type of Physical Defense Mode in use.
         F10 - Activate emergency Physical Defense Mode. Replaces Magical Defense Mode, if that was active.
     ]]
@@ -66,23 +64,23 @@ function user_setup()
 
         F12 - Update currently equipped gear, and report current status.
         Ctrl-F12 - Cycle Idle Mode.
-        
+
         Will automatically set IdleMode to Idle when Pet becomes Engaged and you are Idle
     ]]
     state.IdleMode:options("Idle", "MasterDT")
 
-    --Various Cycles for the different types of PetModes
+    -- Various Cycles for the different types of PetModes
     state.PetStyleCycleTank = M {"NORMAL", "DD", "MAGIC", "SPAM"}
     state.PetStyleCycleMage = M {"NORMAL", "HEAL", "SUPPORT", "MB", "DD"}
     state.PetStyleCycleDD = M {"NORMAL", "BONE", "SPAM", "OD", "ODACC"}
 
-    --The actual Pet Mode and Pet Style cycles
-    --Default Mode is Tank
+    -- The actual Pet Mode and Pet Style cycles
+    -- Default Mode is Tank
     state.PetModeCycle = M {"TANK", "DD", "MAGE"}
-    --Default Pet Cycle is Tank
+    -- Default Pet Cycle is Tank
     state.PetStyleCycle = state.PetStyleCycleTank
 
-    --Toggles
+    -- Toggles
     --[[
         Alt + E will turn on or off Auto Maneuver
     ]]
@@ -109,7 +107,7 @@ function user_setup()
     ]]
     state.SetFTP = M(false, "Set FTP")
 
-   --[[
+    --[[
         This will hide the entire HUB
         //gs c hub all
     ]]
@@ -136,7 +134,7 @@ function user_setup()
     --[[
         This will toggle the HUB lite mode
         //gs c hub lite
-    ]]  
+    ]]
     state.useLightMode = M(false, "Toggles Lite mode")
 
     --[[
@@ -145,12 +143,12 @@ function user_setup()
     ]]
     state.Keybinds = M(false, "Hide Keybinds")
 
-    --[[ 
-        This will toggle the CP Mode 
-        //gs c toggle CP 
-    ]] 
-    state.CP = M(false, "CP") 
-    CP_CAPE = "Aptitude Mantle +1" 
+    --[[
+        This will toggle the CP Mode
+        //gs c toggle CP
+    ]]
+    state.CP = M(false, "CP")
+    CP_CAPE = "Aptitude Mantle +1"
 
     --[[
         Enter the slots you would lock based on a custom set up.
@@ -160,8 +158,8 @@ function user_setup()
         //gs c toggle customgearlock
         ]]
     state.CustomGearLock = M(false, "Custom Gear Lock")
-    --Example customGearLock = T{"head", "waist"}
-    customGearLock = T{}
+    -- Example customGearLock = T{"head", "waist"}
+    customGearLock = T {}
 
     send_command("bind !f7 gs c cycle PetModeCycle")
     send_command("bind ^f7 gs c cycleback PetModeCycle")
@@ -174,16 +172,15 @@ function user_setup()
     send_command("bind home gs c toggle setftp")
     send_command("bind PAGEUP gs c toggle autodeploy")
     send_command("bind PAGEDOWN gs c hide keybinds")
-    send_command("bind end gs c toggle CP") 
+    send_command("bind end gs c toggle CP")
     send_command("bind = gs c clear")
 
     select_default_macro_book()
 
     -- Adjust the X (horizontal) and Y (vertical) position here to adjust the window
     pos_x = 0
-    pos_y = 0
+    pos_y = 500
     setupTextWindow(pos_x, pos_y)
-    
 end
 
 function file_unload()
@@ -197,17 +194,23 @@ function file_unload()
     send_command("unbind ^`")
     send_command("unbind home")
     send_command("unbind PAGEUP")
-    send_command("unbind PAGEDOWN")       
+    send_command("unbind PAGEDOWN")
     send_command("unbind end")
     send_command("unbind =")
+
 end
 
 function job_setup()
-    include("PUP-LIB.lua")
+    include("PUP-LIB/pup-main.lua")
+
+    -- This is to set the mininmum amount of TP you want your pet to have before equiping TP gear (pet is fighting only)
+    PET_MIN_TP_TO_WEAPONSKILL = 850
+    -- This is the seconds of how long to keep the pet weaponskill gear equipped before reverting to previous set
+    PET_GEAR_WEAPONSKILL_LOCKOUT_TIMER = 5
 end
 
 function init_gear_sets()
-    --Table of Contents
+    -- Table of Contents
     ---Gear Variables
     ---Master Only Sets
     ---Hybrid Only Sets
@@ -217,9 +220,9 @@ function init_gear_sets()
     -------------------------------------------------------------------------
     --  _____                  __      __        _       _     _
     -- / ____|                 \ \    / /       (_)     | |   | |
-    --| |  __  ___  __ _ _ __   \ \  / /_ _ _ __ _  __ _| |__ | | ___  ___
-    --| | |_ |/ _ \/ _` | '__|   \ \/ / _` | '__| |/ _` | '_ \| |/ _ \/ __|
-    --| |__| |  __/ (_| | |       \  / (_| | |  | | (_| | |_) | |  __/\__ \
+    -- | |  __  ___  __ _ _ __   \ \  / /_ _ _ __ _  __ _| |__ | | ___  ___
+    -- | | |_ |/ _ \/ _` | '__|   \ \/ / _` | '__| |/ _` | '_ \| |/ _ \/ __|
+    -- | |__| |  __/ (_| | |       \  / (_| | |  | | (_| | |_) | |  __/\__ \
     -- \_____|\___|\__,_|_|        \/ \__,_|_|  |_|\__,_|_.__/|_|\___||___/
     -------------------------------------------------------------------------
     --[[
@@ -230,8 +233,8 @@ function init_gear_sets()
     Animators.Range = "Animator P II"
     Animators.Melee = "Animator P +1"
 
-    --Adjust to your reforge level
-    --Sets up a Key, Value Pair
+    -- Adjust to your reforge level
+    -- Sets up a Key, Value Pair
     Artifact_Foire = {}
     Artifact_Foire.Head_PRegen = "Foire Taj +1"
     Artifact_Foire.Body_WSD_PTank = "Foire Tobe +1"
@@ -240,11 +243,11 @@ function init_gear_sets()
     Artifact_Foire.Feet_Repair_PMagic = "Foire Babouches +1"
 
     Relic_Pitre = {}
-    Relic_Pitre.Head_PRegen = "Pitre Taj +2" --Enhances Optimization
-    Relic_Pitre.Body_PTP = "Pitre Tobe +2" --Enhances Overdrive
-    Relic_Pitre.Hands_WSD = "Pitre Dastanas +2" --Enhances Fine-Tuning
-    Relic_Pitre.Legs_PMagic = "Pitre Churidars +2" --Enhances Ventriloquy
-    Relic_Pitre.Feet_PMagic = "Pitre Babouches +1" --Role Reversal
+    Relic_Pitre.Head_PRegen = "Pitre Taj +2" -- Enhances Optimization
+    Relic_Pitre.Body_PTP = "Pitre Tobe +2" -- Enhances Overdrive
+    Relic_Pitre.Hands_WSD = "Pitre Dastanas +2" -- Enhances Fine-Tuning
+    Relic_Pitre.Legs_PMagic = "Pitre Churidars +2" -- Enhances Ventriloquy
+    Relic_Pitre.Feet_PMagic = "Pitre Babouches +1" -- Role Reversal
 
     Empy_Karagoz = {}
     Empy_Karagoz.Head_PTPBonus = "Karagoz Capello"
@@ -256,23 +259,13 @@ function init_gear_sets()
     Visucius = {}
     Visucius.PetDT = {
         name = "Visucius's Mantle",
-        augments = {
-            "Pet: Acc.+20 Pet: R.Acc.+20 Pet: Atk.+20 Pet: R.Atk.+20",
-            "Accuracy+20 Attack+20",
-            "Pet: Accuracy+4 Pet: Rng. Acc.+4",
-            'Pet: "Regen"+10',
-            "Pet: Damage taken -5%"
-        }
+        augments = {"Pet: Acc.+20 Pet: R.Acc.+20 Pet: Atk.+20 Pet: R.Atk.+20", "Accuracy+20 Attack+20",
+                    "Pet: Accuracy+4 Pet: Rng. Acc.+4", 'Pet: "Regen"+10', "Pet: Damage taken -5%"}
     }
     Visucius.PetMagic = {
         name = "Visucius's Mantle",
-        augments = {
-            "Pet: Acc.+20 Pet: R.Acc.+20 Pet: Atk.+20 Pet: R.Atk.+20",
-            "Accuracy+20 Attack+20",
-            "Pet: Accuracy+4 Pet: Rng. Acc.+4",
-            'Pet: "Regen"+10',
-            "Pet: Damage taken -5%"
-        }
+        augments = {"Pet: Acc.+20 Pet: R.Acc.+20 Pet: Atk.+20 Pet: R.Atk.+20", "Accuracy+20 Attack+20",
+                    "Pet: Accuracy+4 Pet: Rng. Acc.+4", 'Pet: "Regen"+10', "Pet: Damage taken -5%"}
     }
 
     --------------------------------------------------------------------------------
@@ -285,7 +278,7 @@ function init_gear_sets()
     --                                                  __/ |
     --                                                 |___/
     ---------------------------------------------------------------------------------
-    --This section is best utilized for Master Sets
+    -- This section is best utilized for Master Sets
     --[[
         Will be activated when Pet is not active, otherwise refer to sets.idle.Pet
     ]]
@@ -293,32 +286,45 @@ function init_gear_sets()
 
     -------------------------------------Fastcast
     sets.precast.FC = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     -------------------------------------Midcast
-    sets.midcast = {} --Can be left empty
+    sets.midcast = {} -- Can be left empty
 
     sets.midcast.FastRecast = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     -------------------------------------Kiting
-    sets.Kiting = {feet = "Hermes' Sandals"}
+    sets.Kiting = {
+        feet = "Hermes' Sandals"
+    }
 
     -------------------------------------JA
-    sets.precast.FC.Utsusemi = set_combine(sets.precast.FC, {neck = "Magoraga Beads", body = "Passion Jacket"})
+    sets.precast.FC.Utsusemi = set_combine(sets.precast.FC, {
+        neck = "Magoraga Beads",
+        body = "Passion Jacket"
+    })
 
     -- Precast sets to enhance JAs
     sets.precast.JA = {} -- Can be left empty
 
-    sets.precast.JA["Tactical Switch"] = {feet = Empy_Karagoz.Feet_Tatical}
+    sets.precast.JA["Tactical Switch"] = {
+        feet = Empy_Karagoz.Feet_Tatical
+    }
 
-    sets.precast.JA["Ventriloquy"] = {legs = Relic_Pitre.Legs_PMagic}
+    sets.precast.JA["Ventriloquy"] = {
+        legs = Relic_Pitre.Legs_PMagic
+    }
 
-    sets.precast.JA["Role Reversal"] = {feet = Relic_Pitre.Feet_PMagic}
+    sets.precast.JA["Role Reversal"] = {
+        feet = Relic_Pitre.Feet_PMagic
+    }
 
-    sets.precast.JA["Overdrive"] = {body = Relic_Pitre.Body_PTP}
+    sets.precast.JA["Overdrive"] = {
+        body = Relic_Pitre.Body_PTP
+    }
 
     sets.precast.JA["Repair"] = {
         ammo = "Automat. Oil +3",
@@ -335,15 +341,17 @@ function init_gear_sets()
         ear1 = "Burana Earring"
     }
 
-    sets.precast.JA["Activate"] = {back = "Visucius's Mantle"}
+    sets.precast.JA["Activate"] = {
+        back = "Visucius's Mantle"
+    }
 
     sets.precast.JA["Deus Ex Automata"] = sets.precast.JA["Activate"]
 
     sets.precast.JA["Provoke"] = {}
 
-    --Waltz set (chr and vit)
+    -- Waltz set (chr and vit)
     sets.precast.Waltz = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     sets.precast.Waltz["Healing Waltz"] = {}
@@ -352,7 +360,7 @@ function init_gear_sets()
     -- Weaponskill sets
     -- Default set for any weaponskill that isn't any more specifically defined
     sets.precast.WS = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     -- Specific weaponskill sets.  Uses the base set if an appropriate WSMod version isn't found.
@@ -362,13 +370,9 @@ function init_gear_sets()
 
     sets.precast.WS["Victory Smite"] = set_combine(sets.precast.WS, {})
 
-    sets.precast.WS["Shijin Spiral"] =
-        set_combine(
-        sets.precast.WS, {
-            -- Add your set here
-        }
-
-    )
+    sets.precast.WS["Shijin Spiral"] = set_combine(sets.precast.WS, {
+        -- Add your set here
+    })
 
     sets.precast.WS["Howling Fist"] = set_combine(sets.precast.WS, {})
 
@@ -378,7 +382,7 @@ function init_gear_sets()
         Idle Mode = MasterDT
     ]]
     sets.idle.MasterDT = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     -------------------------------------Engaged
@@ -387,7 +391,7 @@ function init_gear_sets()
         Hybrid Mode = Normal
     ]]
     sets.engaged.Master = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     -------------------------------------Acc
@@ -396,7 +400,7 @@ function init_gear_sets()
         Hybrid Mode = Acc
     ]]
     sets.engaged.Master.Acc = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     -------------------------------------TP
@@ -405,7 +409,7 @@ function init_gear_sets()
         Hybrid Mode = TP
     ]]
     sets.engaged.Master.TP = {
-       -- Add your set here
+        -- Add your set here
     }
 
     -------------------------------------DT
@@ -414,34 +418,7 @@ function init_gear_sets()
         Hybrid Mode = DT
     ]]
     sets.engaged.Master.DT = {
-       -- Add your set here 
-    }
-
-    --[[
-        Buff Active: Aftermath: Lv.1
-        Offense Mode = Master
-        (This becomes the new baseset while Aftermath is active) 
-    ]]
-    sets.engaged.Master["Aftermath: Lv.1"] = {
-        -- Add your set here 
-    }
-    
-    --[[
-        Buff Active: Aftermath: Lv.2
-        Offense Mode = Master
-        (This becomes the new baseset while Aftermath is active) 
-    ]]
-    sets.engaged.Master["Aftermath: Lv.2"] = {
-        -- Add your set here 
-    }
-    
-    --[[
-        Buff Active: Aftermath: Lv.3
-        Offense Mode = Master
-        (This becomes the new baseset while Aftermath is active) 
-    ]]
-    sets.engaged.Master["Aftermath: Lv.3"] = {
-        -- Add your set here 
+        -- Add your set here
     }
 
     ----------------------------------------------------------------------------------
@@ -461,7 +438,7 @@ function init_gear_sets()
         Hybrid Mode = Normal
     ]]
     sets.engaged.MasterPet = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     -------------------------------------Acc
@@ -470,7 +447,7 @@ function init_gear_sets()
         Hybrid Mode = Acc
     ]]
     sets.engaged.MasterPet.Acc = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     -------------------------------------TP
@@ -479,7 +456,7 @@ function init_gear_sets()
         Hybrid Mode = TP
     ]]
     sets.engaged.MasterPet.TP = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     -------------------------------------DT
@@ -488,7 +465,7 @@ function init_gear_sets()
         Hybrid Mode = DT
     ]]
     sets.engaged.MasterPet.DT = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     -------------------------------------Regen
@@ -497,34 +474,7 @@ function init_gear_sets()
         Hybrid Mode = Regen
     ]]
     sets.engaged.MasterPet.Regen = {
-       -- Add your set here 
-    }
-
-    --[[
-        Buff Active: Aftermath: Lv.1
-        Offense Mode = MasterPet
-        (This becomes the new baseset while Aftermath is active) 
-    ]]
-    sets.engaged.MasterPet["Aftermath: Lv.1"] = {
         -- Add your set here
-    }
-    
-    --[[
-        Buff Active: Aftermath: Lv.2
-        Offense Mode = MasterPet
-        (This becomes the new baseset while Aftermath is active) 
-    ]]
-    sets.engaged.MasterPet["Aftermath: Lv.2"] = {
-        -- Add your set here
-    }
-    
-    --[[
-        Buff Active: Aftermath: Lv.3
-        Offense Mode = MasterPet
-        (This becomes the new baseset while Aftermath is active) 
-    ]]
-    sets.engaged.MasterPet["Aftermath: Lv.3"] = {
-        -- Add your set here"
     }
 
     ----------------------------------------------------------------
@@ -540,40 +490,40 @@ function init_gear_sets()
 
     -------------------------------------Magic Midcast
     sets.midcast.Pet = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     sets.midcast.Pet.Cure = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     sets.midcast.Pet["Healing Magic"] = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     sets.midcast.Pet["Elemental Magic"] = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     sets.midcast.Pet["Enfeebling Magic"] = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     sets.midcast.Pet["Dark Magic"] = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     sets.midcast.Pet["Divine Magic"] = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     sets.midcast.Pet["Enhancing Magic"] = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     -------------------------------------Idle
     --[[
-        This set will become default Idle Set when the Pet is Active 
+        This set will become default Idle Set when the Pet is Active
         and sets.idle will be ignored
         Player = Idle and not fighting
         Pet = Idle and not fighting
@@ -581,7 +531,7 @@ function init_gear_sets()
         Idle Mode = Idle
     ]]
     sets.idle.Pet = {
-       -- Add your set here 
+        head = "Heyoka Cap"
     }
 
     --[[
@@ -592,15 +542,15 @@ function init_gear_sets()
         Idle Mode = MasterDT
     ]]
     sets.idle.Pet.MasterDT = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     -------------------------------------Enmity
     sets.pet = {} -- Not Used
 
-    --Equipped automatically
+    -- Equipped automatically
     sets.pet.Enmity = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     --[[
@@ -608,13 +558,13 @@ function init_gear_sets()
         F10 if Physical Defense Mode = PetDT
     ]]
     sets.pet.EmergencyDT = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     -------------------------------------Engaged for Pet Only
     --[[
       For Technical Users - This is layout of below
-      sets.idle[idleScope][state.IdleMode][ Pet[Engaged] ][CustomIdleGroups] 
+      sets.idle[idleScope][state.IdleMode][ Pet[Engaged] ][CustomIdleGroups]
 
       For Non-Technical Users:
       If you the player is not fighting and your pet is fighting the first set that will activate is sets.idle.Pet.Engaged
@@ -625,7 +575,7 @@ function init_gear_sets()
         Hybrid Mode = Normal
     ]]
     sets.idle.Pet.Engaged = {
-       -- Add your set here 
+        head = "Taeon Chapeau"
     }
 
     --[[
@@ -633,7 +583,7 @@ function init_gear_sets()
         Hybrid Mode = Acc
     ]]
     sets.idle.Pet.Engaged.Acc = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     --[[
@@ -641,7 +591,7 @@ function init_gear_sets()
         Hybrid Mode = TP
     ]]
     sets.idle.Pet.Engaged.TP = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     --[[
@@ -649,7 +599,7 @@ function init_gear_sets()
         Hybrid Mode = DT
     ]]
     sets.idle.Pet.Engaged.DT = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     --[[
@@ -657,58 +607,24 @@ function init_gear_sets()
         Hybrid Mode = Regen
     ]]
     sets.idle.Pet.Engaged.Regen = {
-       -- Add your set here 
+        -- Add your set here
     }
 
     --[[
         Idle Mode = Idle
         Hybrid Mode = Ranged
     ]]
-    sets.idle.Pet.Engaged.Ranged =
-        set_combine(
-        sets.idle.Pet.Engaged,
-        {
-            legs = Empy_Karagoz.Legs_Combat
-        }
-    )
-
-    --[[
-        Buff Active: Aftermath: Lv.1
-        Offense Mode = Trusts
-        Player = Engaged
-        (This becomes the new baseset while Aftermath is active) 
-    ]]
-    sets.idle.Pet.Engaged["Aftermath: Lv.1"] = {
-        -- Add your set here
-    }
-    
-    --[[
-        Buff Active: Aftermath: Lv.2
-        Offense Mode = Trusts
-        Player = Engaged
-        (This becomes the new baseset while Aftermath is active) 
-    ]]
-    sets.idle.Pet.Engaged["Aftermath: Lv.2"] = {
-        -- Add your set here
-    }
-    
-    --[[
-        Buff Active: Aftermath: Lv.3
-        Offense Mode = Trusts
-        Player = Engaged
-        (This becomes the new baseset while Aftermath is active) 
-    ]]
-    sets.idle.Pet.Engaged["Aftermath: Lv.3"] = {
-        -- Add your set here
-    }
+    sets.idle.Pet.Engaged.Ranged = set_combine(sets.idle.Pet.Engaged, {
+        legs = Empy_Karagoz.Legs_Combat
+    })
 
     -------------------------------------WS
     --[[
         WSNoFTP is the default weaponskill set used
     ]]
     sets.midcast.Pet.WSNoFTP = {
-        head = Empy_Karagoz.Head_PTPBonus,
-       -- Add your set here
+        head = "Pitre Taj +2"
+        -- Add your set here
     }
 
     --[[
@@ -716,44 +632,32 @@ function init_gear_sets()
         then this set will be equipped
     ]]
     sets.midcast.Pet.WSFTP = {
-        head = Empy_Karagoz.Head_PTPBonus,
-       -- Add your set here
+        head = Empy_Karagoz.Head_PTPBonus
+        -- Add your set here
     }
 
-    --[[
-        Base Weapon Skill Set
-        Used by default if no modifier is found
-    ]]
-    sets.midcast.Pet.WS = {}
+    -- Base set without modifier, uses WSNoFTP by default
+    sets.midcast.Pet.WS = set_combine(sets.midcast.Pet.WSNoFTP, {
+        -- Add your gear here that would be different from sets.midcast.Pet.WSNoFTP
+        head = "Pitre Taj +2"
+    })
 
-    --Chimera Ripper, String Clipper
-    sets.midcast.Pet.WS["STR"] = set_combine(sets.midcast.Pet.WSNoFTP, {})
+    -- Chimera Ripper, String Clipper
+    sets.midcast.Pet.WS["STR"] = set_combine(sets.midcast.Pet.WSNoFTP, {
+        -- Add your gear here that would be different from sets.midcast.Pet.WSNoFTP
+    })
 
     -- Bone crusher, String Shredder
-    sets.midcast.Pet.WS["VIT"] =
-        set_combine(
-        sets.midcast.Pet.WSNoFTP,
-        {
-            -- Add your gear here that would be different from sets.midcast.Pet.WSNoFTP
-            head = Empy_Karagoz.Head_PTPBonus
-        }
-    )
+    sets.midcast.Pet.WS["VIT"] = set_combine(sets.midcast.Pet.WSNoFTP, {
+        -- Add your gear here that would be different from sets.midcast.Pet.WSNoFTP
+        head = Empy_Karagoz.Head_PTPBonus
+    })
 
     -- Cannibal Blade
     sets.midcast.Pet.WS["MND"] = set_combine(sets.midcast.Pet.WSNoFTP, {})
 
     -- Armor Piercer, Armor Shatterer
     sets.midcast.Pet.WS["DEX"] = set_combine(sets.midcast.Pet.WSNoFTP, {})
-
-    -- Arcuballista, Daze
-    sets.midcast.Pet.WS["DEXFTP"] =
-        set_combine(
-        sets.midcast.Pet.WSFTP,
-        {
-            -- Add your gear here that would be different from sets.midcast.Pet.WSFTP
-            head = Empy_Karagoz.Head_PTPBonus
-        }
-    )
 
     ---------------------------------------------
     --  __  __ _             _____      _
@@ -765,12 +669,12 @@ function init_gear_sets()
     ---------------------------------------------
     -- Town Set
     sets.idle.Town = {
-       -- Add your set here
+        -- Add your set here
     }
 
     -- Resting sets
     sets.resting = {
-       -- Add your set here
+        -- Add your set here
     }
 
     sets.defense.MasterDT = sets.idle.MasterDT
