+++ README.md
@@ -1,55 +1,47 @@
-# Puppet Gearswap for Windower
+# Recent Changes
+The entire structure is being reworked. I have started to pull out various functions into their own section within the PUP-LIB folder to help better maintain these various functions.
 
-- [Puppet Gearswap for Windower](#puppet-gearswap-for-windower)
-- [Requirements](#requirements)
-- [Features](#features)
-  - [Window Position](#window-position)
-  - [Offense Mode](#offense-mode)
-  - [Hybrid Mode](#hybrid-mode)
-  - [Defense Mode](#defense-mode)
-  - [Kiting Mode](#kiting-mode)
-  - [Idle Mode](#idle-mode)
-  - [Lock Weapon](#lock-weapon)
-  - [Auto Maneuvers Toggle](#auto-maneuvers-toggle)
-  - [Emergency Lock Pet DT Set](#emergency-lock-pet-dt-set)
-  - [Predict](#predict)
-  - [Pet Mode and Pet Styles](#pet-mode-and-pet-styles)
-      - [Current Styles](#current-styles)
-  - [SET FTP](#set-ftp)
-  - [Hide Window](#hide-window)
-  - [Notes On Pet Equipping Weaponskill Gear](#notes-on-pet-equipping-weaponskill-gear)
-  - [Notes on Equipping Enmity Gear](#notes-on-equipping-enmity-gear)
-  - [TODO](#todo)
+To Install: Move the older version out of the Gearswap Data Folder. Then move the new PUP.lua and the PUP-LIB folder (with all of its contents) in the Gearswap Data Folder.
 
-This has become much more stable. Gearsets are in a place where they will not likely change. 
+## Gearswap GUI
+Under Pet Skills the two have been renamed:
+- `Maneuver Queue` is now `Maneuvers Maintained`
+- `Current Queue` is now `Maneuvers To Recast`
 
-Most of the testing now involves ensuring the most flexible gear swapping in at the correct moment for various situations a Puppetmaster may find themselves in.
+### Upcoming
+- Work on the Lite Version - Better support for screen sizes
+- Reworking the Mode section to reflect the exact mode you are in at the time (should allow for better visibility into the exact set you should have on) -- This may or may not work, still figuring out
 
-Please keep in mind that this **currently** a work in progress and any feed back you can provide will be helpful in improving this gearswap for puppets everywhere.
+## Pet Auto Equip Weaponskill Gear
+This has been completely reworked. This now properly locks from changing gear once threshold is hit and allows customizing.
 
-# Requirements
-Mote-Libs is used for this gearswap. By default Windower/Gearswap provides this within its lib folder.
+This will only activate when the:
+    - PetMode is DD
+    - PetStyle is Spam, DD or Bone (regardless of PetMode)
+    - Player is not fighting or Offense Mode is Trusts
 
-[Jump To Table of Contents](#puppet-gearswap-for-windower)
+Two new defaults have been added and are changeable within the PUP.lua under job_setup() (don't have to go into the PUP-LIB at all for this):
+- `PET_MIN_TP_TO_WEAPONSKILL`
+    - This is to set the mininmum amount of TP you want your pet to have before equiping TP gear (pet is fighting only)
 
-# Features
+- `PET_GEAR_WEAPONSKILL_LOCKOUT_TIMER`
+    - This is the seconds of how long to keep the pet weaponskill gear equipped before reverting to previous set
 
-## Window Position
-To adjust the window position you change the values in job_setup
+Two new commands have been introduced (this allows you to change the defauls temporarily)
+- `//gs c wstimer 5`
+    - This allows you the pass in the seconds that you will stay in the pet weaponskill gear while waiting on pet
+- `//gs c tpmin 850`
+    - This is when we will start the Pet Weaponskill Timer and first change into Pet Weapon Skill Gear
 
-    Adjust the X (horizontal) and Y (vertical) position here to adjust the window
-    
-    Default:
-    setupTextWindow(0, 0)  
-    
-    Example:
-    setupTextWindow(1400, 600)
+## Auto Maneuver
+A few extra checks have been added to try and prevent any looping of maneuvers.
 
-The window is now draggable, but it does not currently save the settings of the window.
+### Upcoming
+I am adding in a debuff tracker so we don't attempt to recast while under an effect that would prevent you from casting to start with.
 
-## Offense Mode
-These are for when you are fighting with or without Pet
-When you are IDLE and Pet is ENGAGED that is handled by the Idle Mode
+## Other
+Pet Predict will no longer happen when you press F12 - your gear will still reset
+- This can still be used with `//gs c predict`
 
 Cycle Offense Mode (the offensive half of all 'hybrid' melee modes).
 - `F9`
@@ -84,14 +76,14 @@ Turns off any emergency defense mode.
 [Jump To Table of Contents](#puppet-gearswap-for-windower)
 
 ## Kiting Mode
-- `Alt+F10` 
+- `Alt+F10`
 - `//gs c toggle Kiting`
 
 [Jump To Table of Contents](#puppet-gearswap-for-windower)
 
 ## Idle Mode
 Defaults to PetDT for when set to Idle
-        
+
 Will set IdleMode to Idle when Pet becomes Engaged and you are Idle
 
 So, if you wish to go into a fight in MasterDT when first approaching to get Pet Engaged it will auto switch
@@ -122,14 +114,14 @@ This allows the user to lock the DT set for the pet in place and block all other
 - `//gs c toggle LockPetDT`
 
 [Jump To Table of Contents](#puppet-gearswap-for-windower)
-  
+
 ## Predict
 This will attempt to determine the currently equipped puppet and adjust the Pet Mode and Pet Style.
-- `Alt+F6` 
+- `Alt+F6`
 - `//gs c predict`
 
 [Jump To Table of Contents](#puppet-gearswap-for-windower)
- 
+
 ## Pet Mode and Pet Styles
 This will change the mode of the pet and the style of the pet.
 
@@ -150,19 +142,19 @@ These are the current styles that each Pet Mode contains:
 | Mage  | Normal | Heal  | Support | MB    | DD    |
 
 Cycles forward on Pet Modes
-- `ALT+F7` 
+- `ALT+F7`
 - `//gs c cycle PetModeCycle`
 
-Cycles back on Pet Modes 
+Cycles back on Pet Modes
 - `CTRL+F7`
 - `//gs c cycleback PetModecycle`
 
 Cycles forward on Pet Styles
-- `ALT+F8` 
+- `ALT+F8`
 - `//gs c cycle PetStyleCycle`
 
 Cycles backward on Pet Styles
-- `CTRL+F8` 
+- `CTRL+F8`
 - `//gs c cycleback PetStyleCycle`
 
 [Jump To Table of Contents](#puppet-gearswap-for-windower)
@@ -178,7 +170,7 @@ You can hide the entire HUB Window by using the below command
 - `//gs c hub all`
 
 You can hide the State from the Window by using the below command:
-- `//gs c hub state` 
+- `//gs c hub state`
 
 You can hide the Mode from the Window by using the below command
 - `//gs c hub mode`
@@ -226,3 +218,7 @@ One item to be aware of is that this currenly reads the chat log for when the pe
 - [ ] Add the ability to change Attachments with AutoControl Windower Addon
 
 [Jump To Table of Contents](#puppet-gearswap-for-windower)
+### Upcoming
+- Add support for Aftermath Sets
+- Investigate logic around Enmity Sets - I want to make sure we don't stay in this to long, it may be better to miss a swap than stay in this for to long, may need to look into tracking debuffs on pet
+- Investigate leveraging Packets over watching Chat for certain items
